Générateur de contact CorpsRigide (Convex)

+++ VisualStudio-Math_Physique/VisualStudio-Math_Physique/src/engine/corps-rigide/contact/DonneeContact.cpp
@@ -1,6 +1,6 @@
 #include "DonneeContact.h"
 
-DonneeContact::DonneeContact(CorpsRigide* particules[2], const float& restitution, const float& penetration, const Vecteur3D& pointContact, const Vecteur3D& normalContact)
+DonneeContact::DonneeContact(CorpsRigide* particules[2], const double& restitution, const double& penetration, const Vecteur3D& pointContact, const Vecteur3D& normalContact)
 	: restitution(restitution), penetration(penetration), pointContact(pointContact), normalContact(normalContact)
 {
 	this->corpsRigides[0] = corpsRigides[0];
